+++ packages/infinity-modules/production-log/src/components/AssignRejections.vue
@@ -308,7 +308,9 @@ export default {
         this.updateHourlyStats(payload, qty, data, id);
         this.updateShiftStats(qty);
         this.$nextTick(() => {
-          this.$refs.form[0].reset();
+          this.$refs.form.forEach((f) => {
+            f.reset();
+          });
         });
       }
       this.saving = false;