+++ packages/infinity-modules/production-planning/src/components/actions/EditPlan.vue
@@ -171,7 +171,7 @@
                   <v-col cols="12" sm="4">
                     <validation-provider
                       name="activecavity"
-                      :rules="`required|numeric|max_value:${plan.cavity}|min_value:1`"
+                      :rules="`required|max_value:${plan.cavity}|min_value:0.1`"
                       #default="{ errors }"
                     >
                       <v-text-field
@@ -318,7 +318,7 @@
                           <validation-provider
                             name="familyactivecavity"
                             :vid="`cavity-${item._id}`"
-                            :rules="`required|numeric|max_value:${item.cavity}|min_value:1`"
+                            :rules="`required|max_value:${item.cavity}|min_value:0.1`"
                             #default="{ errors }"
                           >
                             <v-text-field