+++ packages/services/api/site.service.js
@@ -9,6 +9,34 @@ class SiteService {
     return this.request.get(`/server/sites/${siteId}/onboardingcompleted/true`);
   }
 
+  addSmsProvider(payload) {
+    return this.request.post('/server/smsprovider', payload);
+  }
+
+  addWebapp(payload) {
+    return this.request.post('/server/webapp', payload);
+  }
+
+  addReport(payload) {
+    return this.request.post('/server/addreport', payload);
+  }
+
+  addReportView(payload) {
+    return this.request.post('/server/reportview', payload);
+  }
+
+  addReportViewMapping(payload) {
+    return this.request.post('/server/reportviewmapping', payload);
+  }
+
+  addInsightView(payload) {
+    return this.request.post('/server/insightview', payload);
+  }
+
+  addLicense(payload) {
+    return this.request.post('/server/customerlicense', payload);
+  }
+
   getLicense(customerId, siteId) {
     return this.request.get(`/server/customerlicense/${customerId}/${siteId}`);
   }
@@ -31,6 +59,14 @@ class SiteService {
 
   getReports() {
     return this.request.get('/server/reports?withFilters=false');
+  }
+
+  getWebApps() {
+    return this.request.get('/server/webapp');
+  }
+
+  getInsights() {
+    return this.request.get('/server/insightview');
   }
 }